requirements.txt
# Core
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2

# Deep Learning
torch==2.3.1
torchvision==0.18.1

# Image Processing
Pillow==10.3.0
opencv-python==4.9.0.80

# Streamlit App
streamlit==1.33.0

# Utilities
tqdm==4.66.4
joblib==1.4.2

pydantic==2.7.1


